% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNearbyGenes.R
\name{addDistNearestTSS}
\alias{addDistNearestTSS}
\title{Calculate the distance between probe and gene TSS}
\usage{
addDistNearestTSS(data, NearGenes, genome, met.platform, cores = 1)
}
\arguments{
\item{data}{A multi Assay Experiment with both DNA methylation and gene Expression objects}

\item{NearGenes}{A list or a data frame with the pairs gene probes}

\item{genome}{Which genome build will be used: hg38 (default) or hg19.}

\item{met.platform}{DNA methyaltion platform to retrieve data from: EPIC or 450K (default)}

\item{cores}{Number fo cores to be used. Deafult: 1}
}
\description{
Calculate the distance between probe and gene TSS
}
\examples{
\dontrun{
 data <- ELMER:::getdata("elmer.data.example")
  NearbyGenes <- GetNearGenes(data = data, 
                              probes = c("cg15924102", "cg24741609"),  
                              numFlankingGenes = 20)
  NearbyGenes <- addDistNearestTSS(data,NearbyGenes)
  NearbyGenes <- addDistNearestTSS(data,NearbyGenes[[1]])
}
}
