% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Heatmap.Func.R
\name{AddLegend}
\alias{AddLegend}
\title{add lengend}
\usage{
AddLegend(Labels = NULL, cols = NULL, margins = c(1, 1, 1, 1),
  lab.las = 2, cexlab = 0.2 + 1/log10(length(cols)))
}
\arguments{
\item{Labels}{A vector of characters}

\item{cols}{A vector of colors for each characters in Labels}

\item{...}{parameters in image function.}
}
\value{
A legend
}
\description{
add lengend
}

