% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Small.R
\name{Binary}
\alias{Binary}
\title{binary data}
\usage{
Binary(x, Break = 0.3, Break2 = NULL)
}
\arguments{
\item{x}{A matrix.}

\item{Break}{A value to binarize the data.}

\item{Break2}{A value to cut value to 3 categories.}
}
\value{
A binarized matrix.
}
\description{
binary data
}

