% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Small.R
\name{calculateEnrichement}
\alias{calculateEnrichement}
\title{Calculate motif Erichment}
\usage{
calculateEnrichement(foreground, background)
}
\arguments{
\item{foreground}{A nsparseMatrix object in each 1 means the motif is found in a region, 0 not.}

\item{background}{A nsparseMatrix object in each 1 means the motif is found in a region, 0 not.}
}
\description{
Calculates fisher exact test
}
\examples{
foreground <- Matrix::Matrix(sample(0:1,size = 100,replace = TRUE), 
                             nrow = 10, ncol = 10,sparse = TRUE)
rownames(foreground) <- paste0("region",1:10)
colnames(foreground) <- paste0("motif",1:10)
background <- Matrix::Matrix(sample(0:1,size = 100,replace = TRUE), 
                             nrow = 10, ncol = 10,sparse = TRUE)
rownames(background) <- paste0("region",1:10)
colnames(background) <- paste0("motif",1:10)
calculateEnrichement(foreground,background)
}
