% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{chat_console}
\alias{chat_console}
\alias{chat_browser}
\title{Open an interactive chat application}
\usage{
chat_console(chat, quiet = FALSE)

chat_browser(chat, quiet = FALSE)
}
\arguments{
\item{chat}{A chat object created by \code{\link[=new_chat_openai]{new_chat_openai()}} or friends.}

\item{quiet}{If \code{TRUE}, suppresses the initial message that explains how
to use the console.}
}
\value{
(Invisibly) The input \code{chat}.
}
\description{
\itemize{
\item \code{chat_console()} lets you chat interactively in the console.
\item \code{chat_browser()} lets you chat interactively in a browser.
}

Note that these functions will mutate the input \code{chat} object as
you chat because your messages will be appended to the history.
}
