% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{chatApp}
\alias{chatApp}
\title{Open an interactive chat application}
\usage{
chatApp(chat)
}
\arguments{
\item{chat}{A chat object created by \code{\link[=new_chat_openai]{new_chat_openai()}} or friends.}
}
\value{
A \link[shiny:shinyApp]{shiny::shinyApp}.
}
\description{
Use \code{chatApp()} to launch an interactive shiny app for any chat instance.
}
