% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Heatmap.Func.R
\name{cluster.main}
\alias{cluster.main}
\title{cluster functions}
\usage{
cluster.main(x, Rowv = TRUE, Colv = TRUE, distfun = dist,
  distMethod = "euclidean", hclustfun = fastcluster::hclust,
  hclustMethod = "complete", Distance.row = NULL, Distance.col = NULL)
}
\arguments{
\item{x}{A matrix}

\item{Rowv}{A boolean determines if the row dendrogram should be computed and reordered.}

\item{Colv}{A boolean determines if the column dendrogram should be computed and reordered.}

\item{distfun}{function used to compute the distance (dissimilarity) between both rows and columns. Defaults to dist.}

\item{distMethod}{A character to specify method for computing distance. Default to euclidean. See detail for other methods.}

\item{hclustfun}{function used to compute the hierarchical clustering when Rowv or Colv are not dendrograms. Defaults to fastcluster::hclust. Should take as argument a result of distfun and return an object to which as.dendrogram can be applied.}

\item{hclustMethod}{A character to specify method for computing clustering. Default to complete}

\item{Distance.row}{A vector of distance value for rows. If Distance.row was specified, distance calculation step will be skiped for rows.}

\item{Distance.col}{A vector of distance value for columns. If Distance.col was specified, distance calculation step will be skiped for columns.}
}
\value{
A list contains: x the original matrix; rowInd order of row after clustering; ddr dendrograms for rows; colInd order of columns after clustering; ddc dendrograms for columns.
}
\description{
cluster functions
}
\details{
distMethod euclidean: Usual square distance between the two vectors (2 norm).maximum: Maximum distance between two components of x and y (supremum norm).manhattan: Absolute distance between the two vectors (1 norm). canberra: sum(|x_i - y_i| / |x_i + y_i|). Terms with zero numerator and denominator are omitted from the sum and treated as if the values were missing. This is intended for non-negative values (e.g. counts): taking the absolute value of the denominator is a 1998 R modification to avoid negative distances.binary: (aka asymmetric binary): The vectors are regarded as binary bits, so non-zero elements are ‘on’ and zero elements are ‘off’. The distance is the proportion of bits in which only one is on amongst those in which at least one is on. minkowski: The p norm, the pth root of the sum of the pth powers of the differences of the components.
}

