% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{Content}
\alias{Content}
\alias{ContentText}
\alias{ContentImage}
\alias{ContentImageRemote}
\alias{ContentImageInline}
\alias{ContentToolRequest}
\alias{ContentToolResult}
\title{Content types}
\usage{
Content()

ContentText(text = class_missing)

ContentImage()

ContentImageRemote(url = class_missing)

ContentImageInline(type = class_missing, data = class_missing)

ContentToolRequest(
  id = class_missing,
  name = class_missing,
  arguments = class_missing
)

ContentToolResult(
  id = class_missing,
  value = class_missing,
  error = class_missing
)
}
\arguments{
\item{text}{A single string.}

\item{url}{URL to a remote image.}

\item{type}{MIME type of the image.}

\item{data}{Base64 encoded image data.}

\item{id}{Tool call id (used to associate a request and a result)}

\item{name}{Function name}

\item{arguments}{Named list of arguments to call the function with.}

\item{value, error}{Either the results of calling the function if
it succeeded, otherwise the error message, as a string. One of
\code{value} and \code{error} will always be \code{NULL}.}
}
\description{
elmer abstracts away differences in the way that different providers
represent different types of content, so that you can more easily write
code that works with any chatbot.

This set of classes represents the various types of content that can be
sent to and recieved from a chat bot:
\itemize{
\item \code{ContentText}: simple text (often in markdown format). This is the only
type of content that can be streamed live as it's received.
\item \code{ContentImageRemote} and \code{ContentImageInline}: images, either as a pointer
to a remote URL or included inline the object. See \code{\link[=content_image_file]{content_image_file()}}
and friends for convenient ways to construct these objects.
\item \code{ContentToolRequest}: a request to perform a tool call (sent by the
assistant).
\item \code{ContentToolResult}: the result of calling the tool (sent by the user).
}
}
