% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Small.R
\name{createMotifRelevantTfs}
\alias{createMotifRelevantTfs}
\title{Get family of transcription factors}
\usage{
createMotifRelevantTfs(classification = "family")
}
\arguments{
\item{classification}{Select if we will use Family classification or sub-family}
}
\value{
A list of TFs and its family members
}
\description{
This function will use TF Class database to create the object
that maps for each TF the members of its family. TF in the same family have 
high correlared PWM.
}
