% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_pipe.R
\name{createSummaryDocument}
\alias{createSummaryDocument}
\title{Create summary document for TCGA.pipe function}
\usage{
createSummaryDocument(analysis = "all", argument.values = "defaults",
  genome = NULL, mae.path = NULL, direction = NULL, group.col = NULL,
  group1 = NULL, group2 = NULL, results.path = NULL)
}
\description{
This function will create a text file with the 
date of the last run, which aanalysis were performed, the values of
the arguments so the user can keep track
}
