% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELMER.R
\docType{data}
\name{data}
\alias{data}
\title{A MultiAssayExperiment containing
DNA methylation data: 101 probes from platform 450K 
Gene Expression data: 1026 genes
for 234 samples from TCGA-LUSC.
This data is used in the examples of ELMER package}
\format{A MultiAssayExperiment for 234 Samples (8 normal samples, 226 Primary solid tumor)}
\description{
A MultiAssayExperiment containing
DNA methylation data: 101 probes from platform 450K 
Gene Expression data: 1026 genes
for 234 samples from TCGA-LUSC.
This data is used in the examples of ELMER package
}
\keyword{internal}
