% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Heatmap.Func.R
\name{dendro.plot}
\alias{dendro.plot}
\title{Plot the dendro tree.}
\usage{
dendro.plot(Rdend = NULL, Cdend = NULL, margin = c(0.5, 0.5, 0.5, 0.5),
  title = NULL, cex.title)
}
\arguments{
\item{Rdend}{dendrograms for rows.}

\item{Cdend}{dendrograms for columns.}

\item{title}{The main title (on top).}

\item{cex.title}{A numerical value giving the amount by which plotting title text and symbols should be magnified relative to the default. This starts as 1 when a device is opened, and is reset when the layout is changed.}
}
\value{
a graph of dendrograms tree.
}
\description{
Plot the dendro tree.
}

