% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ReadFile.R
\name{Distal}
\alias{Distal}
\title{Generate random loci of genome.}
\usage{
Distal(x, TSS.range, ignore.strand = F)
}
\arguments{
\item{x}{GRange object which you want to identify distal elements from.}

\item{TSS.range}{GRange object which contain promoter infomation}

\item{ignore.strand}{A boolean which to specific if ignore strand or not.}
}
\value{
GRange object which contains elements from x that doesn't overlap with TSS.range.
}
\description{
Generate random loci of genome.
}

