% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellmer-package.R
\docType{package}
\name{ellmer-package}
\alias{ellmer}
\alias{ellmer-package}
\title{ellmer: Call Large Language Model 'API's}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A consistent interface for calling 'LLM' (large language model) APIs that works across different providers including 'Claude', 'OpenAI' 'Azure' 'Bedrock' and 'Google' 'Gemini'. Supports streaming, 'async', tool calling, and structured data extraction.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ellmer.tidyverse.org}
  \item \url{https://github.com/tidyverse/ellmer}
  \item Report bugs at \url{https://github.com/tidyverse/ellmer/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@posit.co}

Authors:
\itemize{
  \item Joe Cheng
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
