% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elmer-package.R
\docType{package}
\name{elmer-package}
\alias{elmer}
\alias{elmer-package}
\title{elmer: Call LLM APIs from R}
\description{
A consistent interface for calling LLM APIs. Includes support for streaming.
}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
