% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELMER.R
\docType{package}
\name{ELMER}
\alias{ELMER}
\title{ELMER is designed to use DNA methylation and gene expression from a
large number of samples to infere regulatory element landscape and transcription
factor network in primary tissue.}
\description{
ELMER is designed to use DNA methylation and gene expression from a
large number of samples to infere regulatory element landscape and transcription
factor network in primary tissue.
}
