% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELMER.R
\docType{data}
\name{GeneExp}
\alias{GeneExp}
\title{A matrix containing gene expression data from TCGA
Gene Expression data: 3842 genes
This data is used in the examples of ELMER package}
\format{A gene expression matrix for 234 Samples and 3842 genes}
\description{
A matrix containing gene expression data from TCGA
Gene Expression data: 3842 genes
This data is used in the examples of ELMER package
}
\keyword{internal}
