% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Main_function.R
\name{get.diff.meth}
\alias{get.diff.meth}
\title{get.diff.meth}
\usage{
get.diff.meth(mee, diff.dir = "both", cores = NULL, percentage = 0.2,
  pvalue = 0.01, sig.dif = 0.3, dir.out = "./", save = TRUE)
}
\arguments{
\item{mee}{A MEE.data object containing at least meth and probeInfo.}

\item{diff.dir}{A character showing differential methylation dirction.
It can be "hypo" which is only selecting hypomethylated probes;
"hyper" which is only selecting hypermethylated probes;
"both" which select both hypomethyalted and hypermethylated probes.}

\item{cores}{A interger which defines number of core to be used in parallel process.
Default is NULL: don't use parallel process.}

\item{percentage}{A number ranges from 0 to 1 specifying the percentage of
samples used to identify the differential methylation. Default is 0.2.}

\item{pvalue}{A number specify the significant Pvalue cutoff for significant
hypo/hyper-methylated probes. Default is current directory.}

\item{sig.dif}{A number specify the significant methylation difference cutoff
for significant hypo/hyper-methylated probes. Default is 0.3.}

\item{dir.out}{A path specify the directory for outputs. Default is is current directory.}

\item{save}{A logic. When TRUE, output file will be saved.}
}
\value{
Statistics for all probes and significant hypo or hyper-methylated probes.
}
\description{
get.diff.meth
}
\examples{
load(system.file("extdata","mee.example.rda",package = "ELMER"))
Hypo.probe <- get.diff.meth(mee, diff.dir="hypo") # get hypomethylated probes
}

