% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Main_function.R
\name{get.enriched.motif}
\alias{get.enriched.motif}
\title{get.enriched.motif}
\usage{
get.enriched.motif(probes.motif, probes, background.probes, lower.OR = 1.1,
  min.incidence = 10, dir.out = "./", label = NULL)
}
\arguments{
\item{probes.motif}{A matrix contains motifs occurrence within probes regions.}

\item{probes}{A vector lists the probes' names in which motif enrichment will be calculated.}

\item{background.probes}{A vector list of probes' names which are considered as
background for motif.enrichment calculation.}

\item{lower.OR}{A number specify the lower boundary of Odds ratio which defines
the significant enriched motif. 1.1 is default.}

\item{min.incidence}{A non-negative integer specify the minimum incidence of
motif in the given probes set. 10 is default.}

\item{dir.out}{A path specify the directory for outputs. Default is current directory}

\item{label}{A character labels the outputs.}
}
\value{
A list contains enriched motifs with the probes regions harboring the motif.
}
\description{
get.enriched.motif
}

