% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Main_function.R
\name{get.feature.probe}
\alias{get.feature.probe}
\title{get.feature.probe}
\usage{
get.feature.probe(feature, TSS, TSS.range = list(upstream = 2000, downstream =
  2000), promoter = FALSE, rm.chr = NULL)
}
\arguments{
\item{feature}{A GRange object containing biofeature coordinate such as
enhancer coordinates. Default is comprehensive genomic enhancer regions from REMC and FANTOM5.
feature option is only usable when promoter option is FALSE.}

\item{TSS}{A GRange object containing the transcription start site. Default is UCSC gene TSS.}

\item{TSS.range}{A list specify how to define promoter regions.
Default is upstream =2000bp and downstream=2000bp.}

\item{promoter}{A logical.If TRUE, function will ouput the promoter probes.
If FALSE, function will ouput the distal probes overlaping with features. The
default is FALSE.}

\item{rm.chr}{A vector of chromosome need to be remove from probes such as chrX chrY or chrM}
}
\value{
A GRange object containing probes that satisfy selecting critiria.
}
\description{
This function selects the probes on HM450K that either overlap
distal biofeatures or TSS promoter.
}
\examples{
# get distal enhancer probe
\dontrun{
Probe <- get.feature.probe()
}
# get promoter probes
\dontrun{
Probe <- get.feature.probe(promoter=FALSE)
}
# get distal enhancer probe remove chrX chrY
Probe2 <- get.feature.probe(rm.chr=c("chrX", "chrY"))
}

