% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticTest.R
\name{Get.Pvalue.p}
\alias{Get.Pvalue.p}
\title{Calculate empirical Pvalue}
\usage{
Get.Pvalue.p(U.matrix, permu)
}
\arguments{
\item{U.matrix}{A data.frame of raw pvalue from U test. Output from .Stat.nonpara}

\item{permu}{data frame of permutation. Output from .Stat.nonpara.permu}
}
\value{
A data frame with empirical Pvalue.
}
\description{
Calculate empirical Pvalue
}

