% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get.tabs}
\alias{get.tabs}
\title{Creating matrix for MR TF heatmap}
\usage{
get.tabs(dir, classification = "family")
}
\arguments{
\item{dir}{Vector ofr directory with results}

\item{classification}{Consider family or subfamily}
}
\description{
Code used to create matrix for MR TF heatmap
}
\examples{
\dontrun{
elmer.results <- dirname(
dir(path = "analysis",
   pattern = "*.hypo.pairs.significant.csv",
   recursive = T,
   full.names = T,
   all.files = T))
tabs <- get.tabs(dir = elmer.results, classification = "subfamily")
}
}
