% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Main_function.R
\name{get.TFs}
\alias{get.TFs}
\title{get.TFs}
\usage{
get.TFs(mee, enriched.motif, TFs, motif.relavent.TFs, percentage = 0.2,
  dir.out = "./", label = NULL, cores = NULL)
}
\arguments{
\item{mee}{A MEE.data object containing at least meth, exp, probeInfo, geneInfo.}

\item{enriched.motif}{Can be either a list containing output of
get.enriched.motif function or path of rda file containing output of get.enriched.motif function.}

\item{TFs}{Can be either a data.frame containing TF GeneID and Symbol
or path of csv file containing TF GeneID and Symbol. If missing,
human TF list will be used. For detail information, refer reference paper.}

\item{motif.relavent.TFs}{Can be either a list containing motif (list name)
and relavent TF (content of list) or path of rda file containing a list
containing motif (list name) and relavent TF (content of list). If missing,
 human TF list will be used. For detail information, refer reference paper.}

\item{percentage}{A number ranges from 0 to 1 specifying the percentage of
samples used to link probes to genes. Default is 0.2.}

\item{dir.out}{A path specify the directory for outputs. Default is current directory}

\item{label}{A character labels the outputs.}

\item{cores}{A interger which defines number of core to be used in parallel
process. Default is NULL: don't use parallel process.}
}
\value{
Potential responsible TFs will be reported.
}
\description{
get.TFs
}
\examples{
load(system.file("extdata","mee.example.rda",package = "ELMER"))
enriched.motif <- list("TP53"= c("cg00329272", "cg10097755", "cg08928189",
                                 "cg17153775", "cg21156590", "cg19749688", "cg12590404",
                                 "cg24517858", "cg00329272", "cg09010107", "cg15386853",
                                 "cg10097755", "cg09247779", "cg09181054"))
TF <- get.TFs(mee, enriched.motif,
TFs=data.frame(GeneID=c("ID7157","ID8626","ID7161"),
              Symbol=c("TP53","TP63","TP73"),
              stringsAsFactors = FALSE),
              label="hypo")
}

