% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchTCGA.R
\name{get450K}
\alias{get450K}
\title{get450K to download HM450K DNA methylation data for certain cancer types
from GDC website.
 @description 
 get450K is a function to download latest version of HM450K DNA methylation 
 for  all samples of certain cancer types from GDC website.}
\usage{
get450K(disease, basedir = "./Data")
}
\arguments{
\item{disease}{A character specifies the disease to download from TCGA such as BLCA}

\item{basedir}{A path. Shows where the data will be stored.}
}
\value{
Download all DNA methylation from HM450K level 3 data for
 the specified disease.
}
\description{
get450K to download HM450K DNA methylation data for certain cancer types
from GDC website.
 @description 
 get450K is a function to download latest version of HM450K DNA methylation 
 for  all samples of certain cancer types from GDC website.
}

