% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/methodAccess.R
\docType{methods}
\name{getExp}
\alias{getExp}
\alias{getExp,MEE.data-method}
\title{getExp}
\usage{
getExp(object, geneID, ID)

\S4method{getExp}{MEE.data}(object, geneID, ID)
}
\arguments{
\item{object}{MEE.data object}

\item{geneID}{A vector of genes' id. When specified, gene expression only 
for these genes will be output.}

\item{ID}{A vector of sample ID. When specified, gene expression only for 
these samples will be output.}
}
\value{
Gene expression values.
}
\description{
getExp
}
\examples{
exp <- matrix(data=c(101:110),ncol=5,dimnames=list(c("gene1","gene2"),paste0("sample",1:5)))
mee <- fetch.mee(exp=exp)
Exp <- getExp(mee, geneID = "gene1") ## get gene expression for certain genes
Exp <- getExp(mee, ID = c("sample1","sample5")) ## get gene expression for certain samples
}

