% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodAccess.R
\name{getExpSamples}
\alias{getExpSamples}
\title{Get Gene expression object samples from MAE}
\usage{
getExpSamples(data)
}
\arguments{
\item{data}{A multiAssayExperiment with DNA methylation and Gene Expression data. See \code{\link{createMAE}} function.}
}
\description{
Get Gene expression object samples from MAE
}
