% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodAccess.R
\name{getMet}
\alias{getMet}
\title{Get DNA methylation object from MAE}
\usage{
getMet(data)
}
\arguments{
\item{data}{A multiAssayExperiment with DNA methylation and Gene Expression data. See \code{\link{createMAE}} function.}
}
\description{
Get DNA methylation object from MAE
}
