% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generic.R, R/methodAccess.R
\docType{methods}
\name{getMeth}
\alias{getMeth}
\alias{getMeth,MEE.data-method}
\title{getMeth}
\usage{
getMeth(object, probe, ID)

\S4method{getMeth}{MEE.data}(object, probe, ID)
}
\arguments{
\item{object}{MEE.data object}

\item{probe}{A vector of probes' name. When specified, DNA methylation only
for these probes will be output.}

\item{ID}{A vector of sample ID. When specified, DNA methylation only for
these samples will be output.}
}
\description{
getMeth
}

