% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/methodAccess.R
\name{getPair}
\alias{getPair}
\title{getPair}
\usage{
getPair(object, probe, geneID)
}
\arguments{
\item{object}{Pair object}

\item{probe}{A vector of probes' name. When specified, only the pair containing
these probes will be output.}

\item{geneID}{A vector of genes' id. When specified, only the pair containing
these genes will be output.}
}
\description{
getPair
}

