% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Small.R
\name{getRegionFromProbes}
\alias{getRegionFromProbes}
\title{Merge nearby probes into regions}
\usage{
getRegionFromProbes(data, pair, distance = 250)
}
\arguments{
\item{data}{Multi Assay Experiment}

\item{pair}{Table with gene-probe pair list}

\item{distance}{Minimum distance to consider from a probe (if it is 100 we will
consider +-100 probe site)}
}
\description{
This function come after the get.pair function.
We will consider nearby probes as a regions controlling the nearby gene
           
0------------  genes
         0-------|
| region |    
Maximum distance between probes will be 1KB
}
\examples{
data(elmer.data.example)
nearGenes <-GetNearGenes(TRange=getMet(data)[c("cg13480549","cg15386853","cg15385853"),],
                         geneAnnot=getExp(data))
Hypo.pair <- get.pair(data = data,
                      nearGenes = nearGenes,
                     permu.size = 5,
                     pvalue = 0.1,
                     Pe = 0.2,
                     dir.out = "./",
                     label = "hypo")
 Hypo.pair <- Hypo.pair[Hypo.pair$GeneID \%in\% Hypo.pair$GeneID[duplicated(Hypo.pair$GeneID)],]
 # No change 
 regions <- getRegionFromProbes(data,Hypo.pair, 1000)  
 # Should be merged as one region
 regions <- getRegionFromProbes(data,Hypo.pair, 3000) 
                   
}
