% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Small.R
\name{getRegionsFromPlatforms}
\alias{getRegionsFromPlatforms}
\title{Transform Probes.motif to regions motif}
\usage{
getRegionsFromPlatforms(probe.gr, probes.motif, distance = 250)
}
\description{
Transform Probes.motif to regions motif
}
\examples{
data("hm450.manifest")
data("Probes.motif.hg19.450K")
regions.motif <- getRegionsFromPlatforms(hm450.manifest,Probes.motif.hg19.450K)
}
