% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generic.R, R/methodAccess.R
\docType{methods}
\name{getSample}
\alias{getSample}
\alias{getSample,MEE.data-method}
\title{getSample}
\usage{
getSample(object, ID, cols)

\S4method{getSample}{MEE.data}(object, ID, cols)
}
\arguments{
\item{object}{MEE.data object}

\item{ID}{A vector of sample ID. When specified, sample informtion only for
these samples will be output.}

\item{cols}{A vector of columns names of Sample slots of MEE.data object.}
}
\description{
getSample
}

