% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Small.R
\name{getSymbol}
\alias{getSymbol}
\title{getSymbol}
\usage{
getSymbol(mee, geneID)
}
\arguments{
\item{mee}{A MEE.data or Pair object.}

\item{geneID}{A character which is the geneID}
}
\value{
gene symbol
}
\description{
getSymbol
}

