% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FetchTCGA.R
\name{getTCGA}
\alias{getTCGA}
\title{getTCGA}
\usage{
getTCGA(disease, Meth = TRUE, RNA = TRUE, Clinic = TRUE,
  basedir = "./Data", RNAtype = "gene", Methfilter = 0.2)
}
\arguments{
\item{disease}{A character to specify disease in TCGA such as BLCA}

\item{Meth}{A logic if TRUE HM450K DNA methylation data will download.}

\item{RNA}{A logic if TRUE RNA-seq Hiseq-V2 from TCGA level 3 will be download.}

\item{Clinic}{A logic if TRUE clinic data will be download for that disease.}

\item{basedir}{A path shows where the data will be stored.}

\item{RNAtype}{A charactor to specify whether use isoform level or gene level. When RNAtype=gene, gene level gene expression will be used. When isoform, then isoform data will be used.}

\item{Methfilter}{A number. For each probe, the percentage of NA among the all the samples should smaller than Methfilter.}
}
\description{
getTCGA
}

