% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Small.R
\name{getTF}
\alias{getTF}
\title{Get human TF list from the UNiprot database}
\usage{
getTF(genome.build = "hg38")
}
\arguments{
\item{genome.build}{Genome reference version "hg38" or "hg19"}
}
\value{
A data frame with the ensemble gene id and entrezgene and gene symbol.
}
\description{
This function gets the last version of human TF list from the UNiprot database
}

