% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Heatmap.Func.R
\name{heatmap.main}
\alias{heatmap.main}
\title{output heatmap}
\usage{
heatmap.main(x, margins = c(5, 0.5, 0.5, 5), labRow = NULL, labCol = NULL,
  nonlab = F, nonlab.row = F, nonlab.col = F, xlab = NULL,
  ylab = NULL, col = heat.colors(225), zlim = NULL, cexRow = 0.2 +
  1/log10(nr), cexCol = 0.2 + 1/log10(nc))
}
\arguments{
\item{x}{output from cluster.main.}

\item{margins}{a character vector of variable names to compute margins for.}

\item{labRow}{a character vector of labels for rows of matrix in x.}

\item{labCol}{a chracter vector of lables for columns of matrix in x.}

\item{nonlab}{a boolean to determine no labels for rows and column.}

\item{nonlab.row}{a boolean to determine no labels for rows.}

\item{nonlab.row}{a boolean to determine no labels for columns.}

\item{...}{parameters for image function.}
}
\value{
A heatmap
}
\description{
output heatmap
}

