% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Heatmap.Func.R
\name{keyplot}
\alias{keyplot}
\title{keyplot}
\usage{
keyplot(x, col, breaks, extremes = NULL, texts, margin = c(1, 1, 1, 1),
  cex.axis = par("cex.axis"))
}
\arguments{
\item{x}{A matrix.}

\item{col}{A vector of colors to define colors.}

\item{breaks}{a vector of number to define how many label in x axis.}

\item{extremes}{a vector of number to define the edge value of the key. Default is NULL.}

\item{...}{parameters for image.}
}
\value{
A keyplot.
}
\description{
keyplot
}

