% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Heatmap.Func.R
\name{lm_eqn}
\alias{lm_eqn}
\title{lable linear regression formula}
\usage{
lm_eqn(df)
}
\arguments{
\item{df}{A data.frame object contains two variables: dependent variable (Dep) and explanation variable (Exp).}
}
\value{
a linear regression formula
}
\description{
lable linear regression formula
}

