% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FetchTCGA.R
\name{matrixClinic}
\alias{matrixClinic}
\title{matrixClinic}
\usage{
matrixClinic(disease, basedir = "./Data")
}
\arguments{
\item{disease}{A character to specify disease in TCGA such as BLCA}

\item{basedir}{A path shows where the data will be stored.}
}
\description{
matrixClinic
}

