% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FetchTCGA.R
\name{matrixMeth}
\alias{matrixMeth}
\title{matrixMeth}
\usage{
matrixMeth(disease, basedir = "./Data", filter = 0.2)
}
\arguments{
\item{disease}{A character to specify disease in TCGA such as BLCA}

\item{basedir}{A path shows where the data will be stored.}

\item{filter}{For each probe, the percentage of NA among the all the samples
should smaller than filter.}
}
\description{
matrixMeth
}

