% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FetchTCGA.R
\name{matrixRNA}
\alias{matrixRNA}
\title{matrixRNA}
\usage{
matrixRNA(disease, basedir = "./Data", type = "gene")
}
\arguments{
\item{disease}{A character to specify disease in TCGA such as BLCA}

\item{basedir}{A path shows where the data will be stored.}

\item{type}{A charactor to specify whether use isoform level or gene level.
When RNAtype=gene, gene level gene expression will be used. When isoform,
then isoform data will be used.}
}
\value{
A matrix of gene expression values for samples (column) and genes (row)
}
\description{
matrixRNA
}

