% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Class.R
\docType{class}
\name{MEE.data-class}
\alias{MEE.data-class}
\title{MEE.data
An S4 class that methylation, expression, sample information, probe
information and gene information.}
\description{
MEE.data
An S4 class that methylation, expression, sample information, probe
information and gene information.
}
\section{Slots}{

\describe{
\item{\code{meth}}{A matrix of DNA methylation. Each row is one probe and each
column is one sample}

\item{\code{exp}}{A matrix of expression. Each row is one gene and each
column is one sample}

\item{\code{sample}}{A data.frame contains sample information}

\item{\code{probeInfo}}{A GRange object contains probe information}

\item{\code{geneInfo}}{A GRange object contains gene information}
}}

