% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELMER.R
\docType{data}
\name{Meth}
\alias{Meth}
\title{A matrix containing DNA methylation beta-values from TCGA
DNA methylation data: 1728 probes
This data is used in the examples of ELMER package}
\format{A MultiAssayExperiment for 268 Samples  and 1728 probes}
\description{
A matrix containing DNA methylation beta-values from TCGA
DNA methylation data: 1728 probes
This data is used in the examples of ELMER package
}
\keyword{internal}
