% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/motif.TF.Plots.R
\name{motif.enrichment.plot}
\alias{motif.enrichment.plot}
\title{motif.enrichment.plot}
\usage{
motif.enrichment.plot(motif.enrichment, significant = NULL, dir.out = "./",
  save = TRUE, label = NULL)
}
\arguments{
\item{motif.enrichment}{A data frame or file path of get.enriched.motif output
motif.enrichment.csv file. See detial for the format of motif.enrichment if a data frame is specified.}

\item{significant}{A list to select subset of motif. Default is NULL. See detail}

\item{dir.out}{A path specify the directory to which the figures will be saved.
Current directory is default.}

\item{save}{A logic. If true (default), figure will be saved to dir.out.}

\item{label}{A character labels the outputs figure.}
}
\value{
A figure shows the enrichment level for selected motifs.
}
\description{
motif.enrichment.plot
}
\details{
motif.enrichment If input data.frame object, it should contain "motif",
"OR", "lowerOR", "upperOR" columns. motif specifies name of motif;
OR specifies Odds Ratio, lowerOR specifies lower boundary of OR (95%) ;
upperOR specifies upper boundary of OR(95%).
}
\examples{
motif.enrichment <- data.frame(motif=c("TP53","NR3C1","E2F1","EBF1","RFX5",
"ZNF143", "CTCF"),
OR=c(19.33,4.83,1, 4.18, 3.67,3.03,2.49),
lowerOR =c(10,3,1.09,1.9,1.5,1.5, 0.82),
upperOR =c(23,5,3,7,6,5,5),
stringsAsFactors=FALSE)
motif.enrichment.plot(motif.enrichment=motif.enrichment,
                     significant=list(OR=3),
                     label="hypo", save=FALSE)
}

