% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Heatmap.Func.R
\name{Normalize}
\alias{Normalize}
\title{Normalization to 0 to 1}
\usage{
Normalize(x, col = FALSE, row = FALSE, na.rm = FALSE)
}
\arguments{
\item{x}{A matrix.}

\item{col}{A boolean to determine normalize by column or not.}

\item{row}{A boolean to determine normalize by row or not.}

\item{na.rm}{A boolean to determine to remove na number or not.}
}
\value{
A normalized matrix.
}
\description{
Normalization to 0 to 1
}

