% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Heatmap.Func.R
\name{NormalizeMedian}
\alias{NormalizeMedian}
\title{Normalization based on median}
\usage{
NormalizeMedian(x, col = FALSE, row = FALSE, na.rm = FALSE)
}
\arguments{
\item{x}{A matrix.}

\item{col}{A boolean to determine normalize by column or not.}

\item{row}{A boolean to determine normalize by row or not.}

\item{na.rm}{A boolean to determine to remove na number or not.}
}
\value{
A normalized matrix.
}
\description{
Normalization based on median
}

