% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Class.R
\docType{class}
\name{Pair-class}
\alias{Pair-class}
\title{An S4 class that pairs information, probe information and gene information.}
\description{
An S4 class that pairs information, probe information and gene information.
}
\section{Slots}{

\describe{
\item{\code{pairInfo}}{A data.frame}

\item{\code{probeInfo}}{A GRanges object.}

\item{\code{geneInfo}}{A GRanges object.}
}}

