% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Heatmap.Func.R
\name{PeakToVenn}
\alias{PeakToVenn}
\title{Making peak sets venn diagram}
\usage{
PeakToVenn(Peaks, ...)
}
\arguments{
\item{Peaks}{A list of Peak sets.}

\item{...}{parameters from VennDiagram package.}
}
\value{
A venn diagram of peaks.
}
\description{
Making peak sets venn diagram
}

