% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Main_function.R
\name{promoterMeth}
\alias{promoterMeth}
\title{promoterMeth}
\usage{
promoterMeth(mee, sig.pvalue = 0.01, percentage = 0.2, save = TRUE)
}
\arguments{
\item{mee}{A MEE.data object must contains meth, exp, probeInfo, geneInfo four component.}

\item{sig.pvalue}{A number specify significant cutoff for gene silenced by promoter
methylation. Default is 0.01}

\item{percentage}{A number ranges from 0 to 1 specifying the percentage of
samples used to link probes to genes. Default is 0.2.}
}
\value{
A data frame contains genes whose expression significantly anti-correlated
with promoter methylation.
}
\description{
promoterMeth
}

