% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{Provider}
\alias{Provider}
\title{A chatbot provider}
\usage{
Provider(base_url = class_missing, extra_args = class_missing)
}
\arguments{
\item{base_url}{The base URL for the API.}

\item{extra_args}{Arbitrary extra arguments to be included in the request body.}
}
\description{
A Provider captures the details of one chatbot service/API. This captures
how the API works, not the details of the underlying large language model.
Different providers might offer the same (open source) model behind a
different API.
}
\details{
To add support for a new backend, you will need to subclass \code{Provider}
(adding any additional fields that your provider needs) and then implement
the various generics that control the behavior of each provider.
}
