% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ReadFile.R
\name{RandomLoci}
\alias{RandomLoci}
\title{Generate random loci of genome.}
\usage{
RandomLoci(SampleSize = NULL, exclusion = NULL, regionWidth = 0)
}
\arguments{
\item{SampleSize}{A number of random loci you want to generate.}

\item{exclusion}{The chromosome you want to exclude such as chrX chrY.}

\item{regionWidth}{The width of each random loci.}
}
\value{
GRange object.
}
\description{
Generate random loci of genome.
}

