% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ReadFile.R
\name{ReadBed}
\alias{ReadBed}
\title{Read a bed file.}
\usage{
ReadBed(x, strand = FALSE, skip = 0, cols = NULL, seqLength = NULL)
}
\arguments{
\item{x}{A path of bed file (characters)}

\item{strand}{A boolean to specific strands. If true, strand column will be filled as input.
If false, strand column will be filled "*""}

\item{skip}{A number to specify how many lines should be removed from bed file.}

\item{cols}{Specify the column to read from bed file.}

\item{seqLength}{Specify custmer seqLength parameter in GRange function}
}
\value{
GRange object containning bed file information.
}
\description{
Read a bed file.
}

