% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ReadFile.R
\name{ReadGFF}
\alias{ReadGFF}
\title{Read a GFF file.}
\usage{
ReadGFF(x, strand = FALSE, skip = 0)
}
\arguments{
\item{x}{A path of GFF file (characters)}

\item{strand}{A boolean to specific strands. If true, strand column will be
filled as input. If false, strand column will be filled "*""}

\item{skip}{A number to specify how many lines should be removed from bed file.}
}
\value{
GRange object containning GFF file information.
}
\description{
Read a GFF file.
}

