% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scatter.plot.R
\name{scatter}
\alias{scatter}
\title{scatter}
\usage{
scatter(meth, exp, legend.title = "Legend", category = NULL, xlab = NULL,
  ylab = NULL, title = NULL, color.value = NULL, lm_line = FALSE)
}
\description{
scatter
}

