% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Schematic.plot.R
\name{schematic.plot}
\alias{schematic.plot}
\title{schematicPlot}
\usage{
schematic.plot(pair, byProbe, byGene, byCoordinate = list(chr = c(), start =
  c(), end = c()), dir.out = "./", save = TRUE, ...)
}
\arguments{
\item{pair}{A pair object. All slots should be included}

\item{byProbe}{A vector of probe names.}

\item{byGene}{A vector of gene ID}

\item{byCoordinate}{A list contains chr, start and end.
byCoordinate=list(chr=c(),start=c(),end=c()).}

\item{dir.out}{A path specify the directory for outputs. Default is current directory}

\item{save}{A logic. If true, figure will be saved to dir.out.}

\item{...}{Parameters for GetNearGenes}
}
\value{
a schematic plot will be produced.
}
\description{
schematicPlot
}
\details{
byProbes When a vector of probes' name are provided,
function will produce schematic plot for each individual probes.
The schematic plot contains probe, nearby 20 (or the number of gene user specified.)
genes and the significantly linked gene with the probe.
}

