% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Schematic_plot.R
\name{schematic}
\alias{schematic}
\title{Generate random loci of genome.}
\usage{
schematic(target.range, Gene.range, special = list(Symbols = c(), colors =
  c()), target.col, fn, save = T)
}
\arguments{
\item{target.range}{A GRange object showing target coordinate.}

\item{Gene.range}{A GRange oject contains coordinate of a list of genes to show on figure.}

\item{special}{A list. Symbols is specific the gene that you want to emphasize and colors to specific the color to each gene you want to emphasize.}

\item{target.col}{A color for target}

\item{save}{Specific save the graph object or not}
}
\value{
A graph object or save as a pdf.
}
\description{
Generate random loci of genome.
}

