% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Heatmap.Func.R
\name{side.bars2}
\alias{side.bars2}
\title{making side bars for heatmap using segment to solve the problem in image that line will very tiny when a lot of samples.}
\usage{
side.bars2(x, side = "colside", order = NULL, margins = c(5, 0.5, 0.5, 5),
  lab = NULL, zlim = NULL, cexlab = 0.2 + 1/log10(nr))
}
\arguments{
\item{x}{A matrix which is same order of the original matrix before cluster. See details}

\item{side}{A character which are either 'colside' or 'rowside' specifying where the side bar locates.}

\item{order}{A vector of number specifying order of side bars.}

\item{...}{parameters for image function.}
}
\value{
a side bar for heatmap
}
\description{
making side bars for heatmap using segment to solve the problem in image that line will very tiny when a lot of samples.
}
\details{
x must be the matrix. If it is colbars, the row number of col bars should be the same as the col number of the matrix. If it is rowbars, the row number of row bars should be the same as the row number of the matrix
}

