% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Small.R
\name{splitmatrix}
\alias{splitmatrix}
\title{splitmatix}
\usage{
splitmatrix(x, by = "row")
}
\arguments{
\item{x}{A matrix}

\item{by}{A character specify if split the matix by row or column.}
}
\value{
A list each of which is the value of each row/column in the matrix.
}
\description{
splitmatix
}

