% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Small.R
\name{standardizeTcgaId}
\alias{standardizeTcgaId}
\title{tcgaSampleType.}
\usage{
standardizeTcgaId(tcgaId)
}
\arguments{
\item{tcgaId}{A vector list TCGA sample barcode.}
}
\value{
Standardized TCGA IDs
}
\description{
tcgaSampleType.
}

