% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/StatisticTest.R
\name{Stat.diff.meth}
\alias{Stat.diff.meth}
\title{Stat.diff.meth}
\usage{
Stat.diff.meth(probe, meths, TN, test = t.test, percentage = 0.2,
  Top.m = NULL)
}
\arguments{
\item{probe}{A charactor specify probe name}

\item{meths}{A matrix contain DNA methylation data.}

\item{TN}{A vector of category of samples.}

\item{test}{A function specify which statistic test will be used.}

\item{percentage}{A number specify the percentage of normal and tumor
samples used in the test.}

\item{Top.m}{A logic. If to identify hypomethylated probe Top.m should be FALSE.
hypermethylated probe is TRUE.}
}
\description{
Stat.diff.meth
}

