% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_function.R
\name{summarizeTF}
\alias{summarizeTF}
\title{Make MR TF binary table}
\usage{
summarizeTF(files = NULL, path = NULL, classification = "family", top = FALSE)
}
\arguments{
\item{files}{Output of get.pairs function: dataframe or file path}

\item{path}{Directory path with the ELMER results. Files with the following pattern
will be selected TF.*with.motif.summary.csv.}

\item{classification}{Consider subfamily or family classifications}

\item{top}{Get only the top potential (default) or all potentials}
}
\description{
This function uses ELMER analysis
results and summarizes the MR TF identified in each analysis
}
