% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TCGA_pipe.R
\name{TCGA.pipe}
\alias{TCGA.pipe}
\title{ELMER analysis pipe for TCGA data.}
\usage{
TCGA.pipe(disease, analysis = "all", wd = "./", cores = NULL,
  Data = NULL, ...)
}
\arguments{
\item{disease}{TCGA short form disease name such as COAD}

\item{analysis}{a vector of characters listing the analysis need to be done.
Analysis are "download","distal.enhancer","diffMeth","pair","motif","TF.search".
Default is "all" meaning all the analysis will be processed.}

\item{wd}{a path showing working dirctory. Default is "./"}

\item{cores}{A interger which defines number of core to be used in parallel process.
Default is NULL: don't use parallel process.}

\item{Data}{A path showing the folder containing DNA methylation, expression and clinic data}

\item{...}{A list of parameters for functions: GetNearGenes, get.feature.probe,
get.diff.meth, get.pair,}
}
\value{
Different analysis results.
}
\description{
ELMER analysis pipe for TCGA data.
}

