% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Small.R
\name{tcgaSampleType}
\alias{tcgaSampleType}
\title{tcgaSampleType.}
\usage{
tcgaSampleType(x)
}
\arguments{
\item{x}{A TCGA sample barcode.}
}
\value{
Tissue type: Tumor, Normal, Control, TRB, cell_line, XP, XCL
}
\description{
tcgaSampleType.
}

