% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELMER.R
\docType{data}
\name{TF.meth.cor}
\alias{TF.meth.cor}
\title{A matrix containing the relashionship between the DNA methylation level at the probes
for a given motif and the gene expression of TF.
This data is used in the examples of ELMER package}
\format{A matrix with 1968 rows (TFs) and 12 columns (motifs)}
\description{
A matrix containing the relashionship between the DNA methylation level at the probes
for a given motif and the gene expression of TF.
This data is used in the examples of ELMER package
}
\keyword{internal}
