% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/motif.TF.Plots.R
\name{TF.rank.plot}
\alias{TF.rank.plot}
\title{TF.rank.plot}
\usage{
TF.rank.plot(motif.pvalue, motif, TF.label, dir.out = "./", save = TRUE)
}
\arguments{
\item{motif.pvalue}{A matrix or a path specifying location of "XXX.with.pvalue.rda" which is output of getTF.}

\item{motif}{A vector of charactor specify the motif to plot}

\item{TF.label}{A list show the label for each motif. If TF.label is not specified, the motif relevant TF and top3 TF will be labeled.}

\item{dir.out}{A path specify the directory to which the figures will be saved. Current directory is default.}

\item{save}{A logic. If true (default), figure will be saved to dir.out.}
}
\description{
TF.rank.plot
}

