% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFsurvival.plot.R
\name{TFsurvival.plot}
\alias{TFsurvival.plot}
\title{Creates survival plot of based on the expression of a TF}
\usage{
TFsurvival.plot(data, TF, xlim = NULL, percentage = 0.3, save = TRUE)
}
\arguments{
\item{data}{A multi assay Experiment with clinical data in the  phenotypic data matrix
containing the following columns: vital_status, days_to_last_follow_up and days_to_death. Default from GDC and TCGAbiolinks}

\item{TF}{A gene symbol}

\item{xlim}{Limit x axis showed in plot}

\item{percentage}{A number ranges from 0 to 1 specifying the percentage of samples in the 
higher and lower expression groups. Default is 0.3}

\item{save}{Save plot as PDF}
}
\description{
This function will create a survival plot for the samples with higher, midium, low expression
of a given transcription factor. 
By defau;t samples with higher expression are the top 30% and the lower expression the bottom 30%.
}
