% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{token_usage}
\alias{token_usage}
\title{Report on token usage in the current session}
\usage{
token_usage()
}
\value{
A data frame
}
\description{
Call this function to find out the cumulative number of tokens that you
have sent and recieved in the current session.
}
\examples{
token_usage()
}
