% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-def.R
\name{tool_arg}
\alias{tool_arg}
\title{Define arguments for a tool}
\usage{
tool_arg(type, description, ..., required = TRUE)
}
\arguments{
\item{type}{Argument type.}

\item{description}{Description of argument as free text.}

\item{...}{Additional JSON Schema properties (e.g. \code{properties}, \code{enum},
\code{pattern}).}

\item{required}{Is the argument required?}
}
\description{
Define arguments for a tool
}
\keyword{internal}
