% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-def.R
\name{ToolDef}
\alias{ToolDef}
\alias{ToolArg}
\title{Define a tool}
\usage{
ToolDef(name, description, arguments = list(), ...)

ToolArg(type, description, required = TRUE, ...)
}
\arguments{
\item{description}{Description of argument as free text.}

\item{...}{Additional JSON Schema properties (e.g. \code{properties}, \code{enum},
\code{pattern}).}

\item{type}{Argument type (\code{"null"}, \code{"boolean"}, \code{"object"}, \code{"array"},
\code{"number"}, or \code{"string"}).}

\item{required}{Is the argument required?}
}
\description{
Define a tool
}
\keyword{internal}
