% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Small.R
\name{txs}
\alias{txs}
\title{txs to fetch USCS gene annotation (transcripts level) from Bioconductor package Homo.sapians. 
If upstream and downstream are specified in TSS list, promoter regions of USCS gene will be generated.}
\usage{
txs(genome.build = "hg19", TSS = list(upstream = NULL, downstream = NULL))
}
\arguments{
\item{genome.build}{Use TxDb.Hsapiens.UCSC.hg38.knownGene instead of TxDb.Hsapiens.UCSC.hg19.knownGene.
Options: hg19 (default) and hg38.}

\item{TSS}{A list. Contains upstream and downstream like TSS=list(upstream, downstream).
When upstream and downstream is specified, coordinates of promoter regions with gene annotation will be generated.}
}
\value{
UCSC gene annotation if TSS is not specified. Coordinates of UCSC gene promoter regions if TSS is specified.
}
\description{
txs is a function to fetch USCS gene annotation (transcripts level) from Bioconductor package Homo.sapians.
If upstream and downstream are specified in TSS list, promoter regions of USCS gene will be generated.
}
\examples{
# get UCSC gene annotation (transcripts level)
\dontrun{
    txs <- txs()
    txs <- txs(genome.build = "hg38")
}
# get coordinate of all UCSC promoter regions +/-1000bp of TSSs
\dontrun{
txs <- txs(TSS=list(upstream=1000, downstream=1000))
}
}
\author{
Lijing Yao (maintainer: lijingya@usc.edu)
}

