% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{type_boolean}
\alias{type_boolean}
\alias{type_integer}
\alias{type_number}
\alias{type_string}
\alias{type_enum}
\alias{type_array}
\alias{type_object}
\title{Type defintions}
\usage{
type_boolean(description = NULL, required = TRUE)

type_integer(description = NULL, required = TRUE)

type_number(description = NULL, required = TRUE)

type_string(description = NULL, required = TRUE)

type_enum(description = NULL, values, required = TRUE)

type_array(description = NULL, items, required = TRUE)

type_object(.description = NULL, ..., .required = TRUE)
}
\arguments{
\item{description, .description}{The purpose of the component. Generally,
the more information that you can provide here, the better.}

\item{required, .required}{Is the component required?}

\item{values}{Character vector of permitted values.}

\item{...}{Name-type pairs defineing the components that the object must
possess.}

\item{item}{The type of the array items. Can be created by any of the
\code{type_} function.}
}
\description{
These functions define object types in a way that chatbots understand and
are used for tool calling and structured data extraction. There names are
based on the \href{https://json-schema.org}{JSON schema}, which is what the APIs
expected behind the scenes, but it's fairly straightforward to translated
from what you know about R.
}
