% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ReadFile.R
\name{WriteBed}
\alias{WriteBed}
\title{Write a bed file from GRange object.}
\usage{
WriteBed(x, save = T, fn = NULL)
}
\arguments{
\item{x}{GRange object}

\item{save}{if save is false, function will return a bed format data.frame. if save is true, fn parameter need to be specific and it output bed file in the path you specified in fn.}

\item{fn}{A name of bed file you want to output.}
}
\value{
A data.frame bed object or save output bed file.
}
\description{
Write a bed file from GRange object.
}

