% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNearbyGenes.R
\name{addDistNearestTSS}
\alias{addDistNearestTSS}
\title{Calculate the distance between probe and gene TSS}
\usage{
addDistNearestTSS(data, NearGenes, cores = 1)
}
\arguments{
\item{data}{A multi Assay Experiment with both DNA methylation and gene Expression objects}

\item{NearGenes}{A list or a data frame with the pairs gene probes}
}
\description{
Calculate the distance between probe and gene TSS
}
\examples{
\dontrun{
 data <- ELMER:::getdata("elmer.data.example")
  NearbyGenes <- GetNearGenes(data = data, probes = c("cg15924102", "cg24741609"),  numFlankingGenes = 20)
  NearbyGenes <- addDistNearestTSS(data,NearbyGenes)
  NearbyGenes <- addDistNearestTSS(data,NearbyGenes[[1]])
}
}
