% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Small.R
\name{createMotifRelevantTfs}
\alias{createMotifRelevantTfs}
\title{Get family of transcription factors}
\usage{
createMotifRelevantTfs(classification = "family")
}
\arguments{
\item{classification}{Select if we will use Family classification or sub-family}
}
\value{
A list of TFs and its family members
}
\description{
This will output a list each TF motif and TFs that binding the motis. Multiple TFs may
recognize a same motif such as TF family.  
The association between each motif famil and transcription factor was created using the 
(HOCOMOCO)[https://hocomoco11.autosome.org/human/mono?full=true] which TF structural families 
was created according to TFClass [@wingender2014tfclass]
This data is stored as a list whose elements 
are motifs and contents for each element are TFs which recognize the same motif that
is the name of the element. This data is used in function get.TFs in \pkg{ELMER} 
to identify the real regulator TF whose motif is enriched in a given set of probes 
and expression associate with average DNA methylation of these motif sites.
}
