% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elmer-package.R
\docType{package}
\name{elmer-package}
\alias{elmer}
\alias{elmer-package}
\title{elmer: Call LLM APIs from R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A consistent interface for calling LLM APIs. Includes support for streaming.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://elmer.tidyverse.org}
  \item \url{https://github.com/tidyverse/elmer}
  \item Report bugs at \url{https://github.com/tidyverse/elmer/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@posit.co}

Authors:
\itemize{
  \item Joe Cheng
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
