% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Small.R
\name{fetch.pair}
\alias{fetch.pair}
\title{fetch.pair}
\usage{
fetch.pair(pair, probeInfo, geneInfo)
}
\arguments{
\item{pair}{A data.frame or path of csv file containing pair information.}

\item{probeInfo}{A GRnage object or path of rda file only containing
a GRange of probe information}

\item{geneInfo}{A GRnage object or path of rda file only containing
a GRange of gene information (Coordinates, GENEID and SYMBOL)}
}
\value{
pair.data object
}
\description{
fetch.pair
}
\examples{
df <- data.frame(Probe=c("cg19403323","cg12213388","cg26607897"),
GeneID =c("ID255928","ID84451","ID55811"),
Symbol =c("SYT14","KIAA1804","ADCY10"),
Pe=c(0.003322259,0.003322259,0.003322259))
geneInfo <- system.file("extdata","UCSC_gene_hg19.rda",package = "ELMER")
## input can be a path
pair <- fetch.pair(pair = df, geneInfo=geneInfo)
}

