% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Small.R
\name{fetch.pair}
\alias{fetch.pair}
\title{fetch.pair}
\usage{
fetch.pair(pair, probeInfo, geneInfo)
}
\arguments{
\item{pair}{A data.frame or path of csv file containing pair information.}

\item{probeInfo}{A GRnage object or path of rda file only containing a GRange of probe information}

\item{geneInfo}{A GRnage object or path of rda file only containing a GRange of gene information (Coordinates, GENEID and SYMBOL)}
}
\value{
pair.data object
}
\description{
fetch.pair
}

