% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_function.R
\name{get.diff.meth}
\alias{get.diff.meth}
\title{get.diff.meth to identify hypo/hyper-methylated CpG sites on HM450K between control and experimental 
groups such as normal verus tumor samples.}
\usage{
get.diff.meth(data, diff.dir = "hypo", cores = 1, percentage = 0.2,
  pvalue = 0.01, group.col, group1, group2, test = t.test, sig.dif = 0.3,
  dir.out = "./", save = TRUE)
}
\arguments{
\item{data}{A multiAssayExperiment with DNA methylation and Gene Expression data. See \code{\link{createMAE function}}.}

\item{diff.dir}{A character can be "hypo" or "hyper", showing dirction DNA methylation changes. If it is "hypo", 
get.diff.meth function will identify all significantly hypomethylated CpG sites; 
If "hyper", get.diff.meth function will identify all significantly hypoermethylated CpG sites}

\item{cores}{A interger which defines the number of cores to be used in parallel process. Default is 1: no parallel process.}

\item{percentage}{A number ranges from 0 to 1 specifying the percentage of samples from control and
experimental groups that are used to identify the differential methylation. Default is 0.2.}

\item{pvalue}{A number specifies the significant P value (adjusted P value by BH) cutoff 
for selecting significant hypo/hyper-methylated probes. Default is 0.01}

\item{group.col}{A column defining the groups of the sample. You can view the 
available columns using: colnames(MultiAssayExperiment::pData(data)).}

\item{test}{Statistical test to be used. Options: t.test (DEFAULT), wilcox.test}

\item{sig.dif}{A number specifies the smallest DNA methylation difference as a cutoff for 
selecting significant hypo/hyper-methylated probes. Default is 0.3.}

\item{dir.out}{A path specify the directory for outputs. Default is is current directory.}

\item{save}{A logic. When TRUE, two getMethdiff.XX.csv files will be generated (see detail)}
}
\value{
Statistics for all probes and significant hypo or hyper-methylated probes.
}
\description{
get.diff.meth applys one-way t-test to identify the CpG sites that are significantly 
hypo/hyper-methyalated using proportional samples (defined by percentage option) from control 
and experimental groups. The P values will be adjusted by Benjamini-Hochberg method. 
Option pvalue and sig.dif will be the criteria (cutoff) for selecting significant differentially methylated CpG sites.
 If save is TURE, two getMethdiff.XX.csv files will be generated (see detail).
}
\details{
save: 
 When save is TRUE, function will generate two XX.csv files.The first one is named 
 getMethdiff.hypo.probes.csv (or getMethdiff.hyper.probes.csv depends on diff.dir). 
 The first file contains all statistic results for each probe. Based on this
 file, user can change different P value or sig.dir cutoff to select the significant results
 without redo the analysis. The second file is named getMethdiff.hypo.probes.significant.csv
 (or getMethdiff.hyper.probes.significant.csv depends on diff.dir). This file contains
 statistic results for the probes that pass the significant criteria (P value and sig.dir).
 When save is FALSE, a data frame R object will be generate which contains the same
 information with the second file.
}
\examples{
data(elmer.data.example)
Hypo.probe <- get.diff.meth(data, 
                            diff.dir="hypo",
                            group.col = "definition", 
                            group1 = "Primary solid Tumor", 
                            group2 = "Solid Tissue Normal",
                            sig.dif = 0.1) # get hypomethylated probes
Hyper.probe <- get.diff.meth(data, 
                            diff.dir="hyper",
                            group.col = "definition", 
                            sig.dif = 0.1) # get hypomethylated probes
}
\references{
Yao, Lijing, et al. "Inferring regulatory element landscapes and transcription 
factor networks from cancer methylomes." Genome biology 16.1 (2015): 1.
}

