% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Main_function.R
\name{get.enriched.motif}
\alias{get.enriched.motif}
\title{get.enriched.motif}
\usage{
get.enriched.motif(probes.motif, probes, background.probes, lower.OR = 1.1,
  min.incidence = 10, dir.out = "./", label = NULL)
}
\arguments{
\item{probes.motif}{A matrix contains motifs occurrence within probes regions.}

\item{probes}{A vector lists the probes' names in which motif enrichment will be calculated.}

\item{background.probes}{A vector list of probes' names which are considered as
background for motif.enrichment calculation.}

\item{lower.OR}{A number specify the lower boundary of Odds ratio which defines
the significant enriched motif. 1.1 is default.}

\item{min.incidence}{A non-negative integer specify the minimum incidence of
motif in the given probes set. 10 is default.}

\item{dir.out}{A path specify the directory for outputs. Default is current directory}

\item{label}{A character labels the outputs.}
}
\value{
A list contains enriched motifs with the probes regions harboring the motif.
}
\description{
get.enriched.motif
}
\examples{
probes <- c("cg00329272","cg10097755","cg08928189", "cg17153775","cg21156590",
"cg19749688","cg12590404","cg24517858","cg00329272","cg09010107",
"cg15386853", "cg10097755", "cg09247779","cg09181054","cg19371916")
load(system.file("extradata","mee.example.rda",package = "ELMER"))
bg <- rownames(getMeth(mee))
enriched.motif <- get.enriched.motif(probes=probes,background.probes = bg,
min.incidence=2, label="hypo")
}

