% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Main_function.R
\name{get.feature.probe}
\alias{get.feature.probe}
\title{get.feature.probe}
\usage{
get.feature.probe(probe, distal = TRUE, feature, TSS,
  TSS.range = list(upstream = 2000, downstream = 2000), rm.chr = NULL)
}
\arguments{
\item{probe}{A GRange object containing probes coordinate information.
Default is Illumina-methyl-450K probes coordinates.}

\item{distal}{A logical. If FALSE, function will output the all
probes overlaping with features. If TRUE, function will ouput the distal probes
overlaping with features.}

\item{feature}{A GRange object containing biofeature coordinate such as
enhancer coordinates. Default is comprehensive genomic enhancer regions from REMC and FANTOM5.}

\item{TSS}{A GRange object containing the transcription start site. Default is UCSC gene TSS.}

\item{TSS.range}{A list specify how to define promoter regions.
Default is upstream =2000bp and downstream=2000bp.}

\item{rm.chr}{A vector of chromosome need to be remove from probes such as chrX chrY or chrM}
}
\value{
A GRange object containing probes that satisfy selecting critiria.
}
\description{
get.feature.probe
}
\examples{
# get distal enhancer probe
Probe <- get.feature.probe()
# get distal enhancer probe which are 1kb away from the promoter regions
Probe1 <- get.feature.probe(TSS.range=list(upstream=1000, downstream=1000))
# get distal enhancer probe remove chrX chrY
Probe2 <- get.feature.probe(rm.chr=c("chrX", "chrY"))
}

