% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_function.R
\name{get.pair}
\alias{get.pair}
\title{get.pair to predict enhancer-gene linkages.}
\usage{
get.pair(data, nearGenes, percentage = 0.2, permu.size = 10000, permu.dir = NULL, 
         pvalue = 0.05, Pe = 0.001, dir.out = "./", calculate.Pe = TRUE, diffExp = FALSE,
         cores = 1, portion = 0.3,  label = NULL, save = TRUE)
}
\arguments{
\item{data}{A multiAssayExperiment with DNA methylation and Gene Expression data. See \code{\link{createMAE}} function.}

\item{nearGenes}{Can be either a list containing output of GetNearGenes 
function or path of rda file containing output of GetNearGenes function.}

\item{percentage}{A number ranges from 0 to 0.5 specifying the percentage of 
samples used to link probes to genes. Default is 0.2.}

\item{permu.size}{A number specify the times of permuation. Default is 10000.}

\item{permu.dir}{A path where the output of permutation will be.}

\item{pvalue}{A number specify the raw p-value cutoff for defining signficant pairs.
Default is 0.05. It will select the significant P value (adjusted P value by BH) cutoff before calculating the empirical p-values.}

\item{Pe}{A number specify the empirical p-value cutoff for defining signficant pairs.
Default is 0.001. Only used if calculate.empirical.p is TRUE.}

\item{dir.out}{A path specify the directory for outputs. Default is current directory}

\item{calculate.Pe}{A logic. If TRUE (default) also execute the permutation step, 
if FALSE only uses the adjusted raw p-value to select the significant pairs.}

\item{diffExp}{A logic. Default is FALSE. If TRUE, t test will be applied to 
test whether putative target gene are differentially expressed between two groups.}

\item{cores}{A interger which defines number of core to be used in parallel process.
Default is 1: don't use parallel process.}

\item{filter.probes}{Should filter probes by selecting only probes that have at least
a certain number of samples below and above a certain cut-off. 
See \code{\link{preAssociationProbeFiltering}} function.}

\item{filter.portion}{A number specify the cut point to define binary methlation level for probe loci. 
Default is 0.3. When beta value is above 0.3, the probe is methylated and 
vice versa. For one probe, the percentage of methylated and unmethylated samples 
should be above filter.percentage value.   
Only used if filter.probes is TRUE. See \code{\link{preAssociationProbeFiltering}} function.}

\item{filter.percentage}{Minimun percentage of samples to be considered in methylated and unmethylated
for the filter.portion option. Default 5\%. Only used if filter.probes is TRUE.
 See \code{\link{preAssociationProbeFiltering}} function.}

\item{label}{A character labels the outputs.}

\item{save}{Two files will be saved if save is true: getPair.XX.all.pairs.statistic.csv
and getPair.XX.pairs.significant.csv (see detail).}
}
\value{
Statistics for all pairs and significant pairs
}
\description{
get.pair is a function to predict enhancer-gene linkages using associations between 
DNA methylation at enhancer CpG sites and expression of 20 nearby genes of the CpG sites
(see reference). Two files will be saved if save is true: getPair.XX.all.pairs.statistic.csv
and getPair.XX.pairs.significant.csv (see detail).
}
\examples{
data(elmer.data.example)
nearGenes <-GetNearGenes(TRange=getMet(data)[c("cg00329272","cg10097755"),],
                         geneAnnot=getExp(data))
Hypo.pair <- get.pair(data=data,
                     nearGenes=nearGenes,
                     permu.size=5,
                     pvalue = 0.2,
                     dir.out="./",
                     label= "hypo")
}
\author{
Lijing Yao (creator: lijingya@usc.edu) 
Tiago C Silva (maintainer: tiagochst@usp.br)
}
\references{
Yao, Lijing, et al. "Inferring regulatory element landscapes and transcription 
factor networks from cancer methylomes." Genome biology 16.1 (2015): 1.
}

