% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_function.R
\name{get.pair}
\alias{get.pair}
\title{get.pair to predict enhancer-gene linkages.}
\usage{
get.pair(data, probes, nearGenes, percentage = 0.2, permu.size = 10000, permu.dir = NULL,  
         Pe = 0.001, dir.out = "./", diffExp = FALSE, cores = 1, portion=0.3,   
         label = NULL, save=TRUE)
}
\arguments{
\item{data}{A multiAssayExperiment with DNA methylation and Gene Expression data. See \code{\link{createMAE function}}.}

\item{nearGenes}{Can be either a list containing output of GetNearGenes 
function or path of rda file containing output of GetNearGenes function.}

\item{percentage}{A number ranges from 0 to 1 specifying the percentage of 
samples used to link probes to genes. Default is 0.2.}

\item{permu.size}{A number specify the times of permuation. Default is 1000.}

\item{permu.dir}{A path where the output of permuation will be.}

\item{Pe}{A number specify the empircal pvalue cutoff for defining signficant pairs.
Default is 0.001
@param portion A number specify the cut point for methylated and unmethylated.
Default is 0.3.
@param diffExp A logic. Default is FALSE. If TRUE, t test will be applied to 
test whether putative target gene are differentially expressed between two groups.}

\item{dir.out}{A path specify the directory for outputs. Default is current directory}

\item{cores}{A interger which defines number of core to be used in parallel process.
Default is 1: don't use parallel process.}

\item{label}{A character labels the outputs.}
}
\value{
Statistics for all pairs and significant pairs
}
\description{
get.pair is a function to predict enhancer-gene linkages using associations between 
DNA methylation at enhancer CpG sites and expression of 20 nearby genes of the CpG sites
(see reference). Two files will be saved if save is true: getPair.XX.all.pairs.statistic.csv
and getPair.XX.pairs.significant.csv (see detail).
}
\examples{
data(elmer.data.example)
nearGenes <-GetNearGenes(TRange=getMet(data)[c("cg00329272","cg10097755"),],
                         geneAnnot=getExp(data))
Hypo.pair <-get.pair(data=data,
                     nearGenes=nearGenes,
                     permu.size=5,
                     Pe = 0.2,
                     dir.out="./",
                     label= "hypo")
}
\author{
Lijing Yao (creator: lijingya@usc.edu) 
Tiago C Silva (maintainer: tiagochst@usp.br)
}
\references{
Yao, Lijing, et al. "Inferring regulatory element landscapes and transcription 
factor networks from cancer methylomes." Genome biology 16.1 (2015): 1.
}

