% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Main_function.R
\name{get.permu}
\alias{get.permu}
\title{get.permu}
\usage{
get.permu(mee, geneID, percentage = 0.2, rm.probes = NULL,
  permu.size = 1000, permu.dir = NULL, cores = NULL)
}
\arguments{
\item{mee}{A MEE.data object containing at least meth, exp, probeInfo, geneInfo.}

\item{geneID}{A vector lists the genes' ID.}

\item{percentage}{A number ranges from 0 to 1 specifying the percentage of samples used to link probes to genes. Default is 0.2.}

\item{rm.probes}{A vector lists the probes name.}

\item{permu.size}{A number specify the times of permuation. Default is 1000.}

\item{permu.dir}{A path where the output of permuation will be.}

\item{cores}{A interger which defines number of core to be used in parallel process. Default is NULL: don't use parallel process.}
}
\value{
permutation
}
\description{
get.permu
}

