% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FetchTCGA.R
\name{get450K}
\alias{get450K}
\title{get450K}
\usage{
get450K(disease, basedir = "./Data")
}
\arguments{
\item{disease}{A character to specify disease in TCGA such as BLCA}

\item{basedir}{A path shows where the data will be stored.}
}
\description{
get450K
}

