% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generic.R, R/methodAccess.R
\docType{methods}
\name{getExp}
\alias{getExp}
\alias{getExp,MEE.data-method}
\title{getExp}
\usage{
getExp(object, geneID, ID)

\S4method{getExp}{MEE.data}(object, geneID, ID)
}
\arguments{
\item{object}{MEE.data object}

\item{geneID}{A vector of genes' id. When specified, gene expression only
for these genes will be output.}

\item{ID}{A vector of sample ID. When specified, gene expression only for
these samples will be output.}
}
\description{
getExp
}

