% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodAccess.R
\name{getExp}
\alias{getExp}
\title{Get Gene expression object from MAE}
\usage{
getExp(data)
}
\arguments{
\item{data}{A multiAssayExperiment with DNA methylation and Gene Expression data. See \code{\link{createMAE}} function.}
}
\description{
Get Gene expression object from MAE
}

