% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/methodAccess.R
\name{getExp}
\alias{getExp}
\title{getExp}
\usage{
getExp(object, geneID, ID)
}
\arguments{
\item{object}{MEE.data object}

\item{geneID}{A vector of genes' id. When specified, gene expression only for these genes will be output.}

\item{ID}{A vector of sample ID. When specified, gene expression only for these samples will be output.}
}
\description{
getExp
}

