% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generic.R, R/methodAccess.R
\docType{methods}
\name{getGeneInfo}
\alias{getGeneInfo}
\alias{getGeneInfo,ANY-method}
\title{getProbeInfo}
\usage{
getGeneInfo(object, geneID, symbol, range)

\S4method{getGeneInfo}{ANY}(object, geneID, symbol, range)
}
\arguments{
\item{object}{MEE.data or Pair object}

\item{geneID}{A vector of genes' id. When specified, only the these genes'
coordinate will be output.}

\item{symbol}{A vector of genes' symbols . When specified, only the these genes'
coordinate will be output.}

\item{range}{A GRanges object. When specified, only the geneInfo locating within
these loci will be output.}
}
\description{
getProbeInfo
}

