% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/methodAccess.R
\name{getMeth}
\alias{getMeth}
\title{getMeth}
\usage{
getMeth(object, probe, ID)
}
\arguments{
\item{object}{MEE.data object}

\item{probe}{A vector of probes' name. When specified, DNA methylation only for these probes will be output.}

\item{ID}{A vector of sample ID. When specified, DNA methylation only for these samples will be output.}
}
\description{
getMeth
}

