% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/methodAccess.R
\docType{methods}
\name{getMeth}
\alias{getMeth}
\alias{getMeth,MEE.data-method}
\title{getMeth}
\usage{
getMeth(object, probe, ID)

\S4method{getMeth}{MEE.data}(object, probe, ID)
}
\arguments{
\item{object}{MEE.data object}

\item{probe}{A vector of probes' name. When specified, DNA methylation only 
for these probes will be output.}

\item{ID}{A vector of sample ID. When specified, DNA methylation only for 
these samples will be output.}
}
\value{
DNA methylation values.
}
\description{
getMeth
}
\examples{
meth <- matrix(data=c(1:20),ncol=5,dimnames=list(paste0("probe",1:4),paste0("sample",1:5)))
mee <- fetch.mee(meth=meth)
Meth <- getMeth(mee,probe = "probe1")
Meth <- getMeth(mee, ID = c("sample1","sample2"))
}

