% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GetNearbyGenes.R
\name{GetNearGenes}
\alias{GetNearGenes}
\title{Collect nearby gene for one locus.}
\usage{
GetNearGenes(geneNum = 20, geneAnnot = NULL, TRange = NULL,
  cores = NULL)
}
\arguments{
\item{geneNum}{A number determine how many gene will be collected from
each side of target (number shoule be even) Default to 20.}

\item{geneAnnot}{A GRange object contains coordinates of promoters for
human genome.}

\item{TRange}{A GRange object contains coordinate of a list targets.}

\item{cores}{A number to specific how many cores to use to compute.
Default to detectCores()/2.}
}
\value{
A data frame of nearby genes and information: genes' IDs, genes' symbols,
distance with target and side to which the gene locate to the target.
}
\description{
Collect nearby gene for one locus.
}
\examples{
load(system.file("extdata","UCSC_gene_hg19.rda",package = "ELMER"))
probe <- GRanges(seqnames = c("chr1","chr2"),
range=IRanges(start = c(16058489,236417627), end= c(16058489,236417627)),
name= c("cg18108049","cg17125141"))
NearbyGenes <- GetNearGenes(geneNum=20,geneAnnot=txs,TRange=probe)
}

