% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNearbyGenes.R
\name{GetNearGenes}
\alias{GetNearGenes}
\title{GetNearGenes to collect nearby genes for one locus.}
\usage{
GetNearGenes(geneAnnot = NULL, TRange = NULL, geneNum = 20, cores = 1)
}
\arguments{
\item{geneAnnot}{A GRange object. Contains coordinates of promoters for 
human genome.}

\item{TRange}{A GRange object. Contains coordinates of a list of targets loci.}

\item{geneNum}{A number determines how many gene will be collected totally. 
Then the number devided by 2 is the number of genes collected from 
each side of targets (number shoule be even) Default to 20.}

\item{cores}{A interger which defines the number of cores to be used in parallel process.
Default is 1: no parallel process.}
}
\value{
A data frame of nearby genes and information: genes' IDs, genes' symbols, 
distance with target and side to which the gene locate to the target.
}
\description{
GetNearGenes is a function to collect equal number of gene on each side of one locus.
}
\examples{
geneAnnot <- txs(TSS=list(upstream=0, downstream=0))
probe <- GRanges(seqnames = c("chr1","chr2"), 
range=IRanges(start = c(16058489,236417627), end= c(16058489,236417627)), 
name= c("cg18108049","cg17125141"))
NearbyGenes <- GetNearGenes(geneNum=20,geneAnnot=geneAnnot,TRange=probe)
}
\references{
Yao, Lijing, et al. "Inferring regulatory element landscapes and transcription 
factor networks from cancer methylomes." Genome biology 16.1 (2015): 1.
}

