% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchTCGA.R
\name{getRNAseq}
\alias{getRNAseq}
\title{getRNAseq to download all RNAseq data for a certain cancer type from TCGA.}
\usage{
getRNAseq(disease, basedir = "./Data", genome = "hg38")
}
\arguments{
\item{disease}{A character specifies disease in TCGA such as BLCA}

\item{basedir}{Download all RNA seq level 3 data for the specified disease.}

\item{genome}{Data aligned against which genome of reference. Options: "hg19", "hg38" (default)}
}
\value{
Download all RNA seq level 3 data for the specified disease.
}
\description{
getRNAseq is a function to download RNAseq data for all samples of a certain cancer type from TCGA
}
