% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Small.R, R/methodAccess.R
\name{getSymbol}
\alias{getSymbol}
\title{getSymbol to report gene symbol from id}
\usage{
getSymbol(data, geneID)

getSymbol(data, geneID)
}
\arguments{
\item{data}{A multiAssayExperiment with DNA methylation and Gene Expression data. See \code{\link{createMAE function}}.}

\item{geneID}{A character which is the ensembl_gene_id}

\item{data}{A multiAssayExperiment with DNA methylation and Gene Expression data. See \code{\link{createMultiAssayExperiment function}}.}

\item{geneID}{A character which is the geneID}
}
\value{
The gene symbol for input genes.

The gene symbol
}
\description{
getSymbol to report gene symbol from id

getSymbol
}
\examples{
data(elmer.data.example)
getSymbol(data, geneID="ENSG00000143067")
}

