% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Small.R
\name{getSymbol}
\alias{getSymbol}
\title{getSymbol}
\usage{
getSymbol(mee, geneID)
}
\arguments{
\item{mee}{A MEE.data or Pair object.}

\item{geneID}{A character which is the geneID}
}
\value{
gene symbol
}
\description{
getSymbol
}
\examples{
geneInfo <- system.file("extdata","UCSC_gene_hg19.rda",package = "ELMER")
## input can be a path
pair <- fetch.pair(geneInfo=geneInfo)
getSymbol(pair, geneID="84451")
}

