% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodAccess.R
\name{getSymbol}
\alias{getSymbol}
\title{getSymbol to report gene symbol from id}
\usage{
getSymbol(data, geneID)
}
\arguments{
\item{data}{A multiAssayExperiment with DNA methylation and Gene Expression data. See \code{\link{createMAE}} function.}

\item{geneID}{A character which is the ensembl_gene_id}
}
\value{
The gene symbol for input genes.
}
\description{
getSymbol to report gene symbol from id
}
\examples{
data <- ELMER:::getdata("elmer.data.example")
getSymbol(data, geneID="ENSG00000143067")
}
