% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Small.R
\name{getTF}
\alias{getTF}
\title{Get human TF list from the UNiprot database}
\usage{
getTF()
}
\value{
A data frame with the ensemble gene id.
}
\description{
This function gets the last version of human TF list from the UNiprot database
}
