% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxPlot.R
\name{heatmapPairs}
\alias{heatmapPairs}
\title{Heatmp  of pairs gene and probes anti-correlated}
\usage{
heatmapPairs(data, group.col, group1, group2, pairs, filename = NULL)
}
\arguments{
\item{data}{A MultiAssayExperiment with a DNA methylation martrix or a DNA methylation matrix}

\item{group1}{A group from group.col. ELMER will run group1 vs group2. 
That means, if direction is hyper, get probes
hypermethylated in group 1 compared to group 2.}

\item{group2}{A group from group.col. ELMER will run group1 vs group2. 
That means, if direction is hyper, get probes
hypermethylated in group 1 compared to group 2.}

\item{pairs}{List of probe and pair genes}

\item{filename}{File names (.pdf) to save the file (i.e. "plot.pdf"). If NULL return plot.}
}
\value{
A heatmap
}
\description{
Heatmp plot of pairs gene and probes anti-correlated
}
\author{
Tiago Chedraoui Silva (tiagochst at gmail.com)
#' @examples
\dontrun{
  data <- ELMER:::getdata("elmer.data.example")
  group.col <- "subtype_Expression.Subtype"
  group1 <- "classical"
  group2 <- "secretory"
  pairs <- data.frame(Probe = c("cg15924102","cg19403323", "cg22396959"),
  GeneID = c("ENSG00000196878", "ENSG00000009790", "ENSG00000009790" ),
  Symbol = c("TRAF3IP3","LAMB3","LAMB3"),
  Raw.p =c(0.001,0.00001,0.001),
  Pe = c(0.001,0.00001,0.001))
  heatmapPairs(data, group.col,group1,group2,pairs,filename = "heatmap.pdf")
  }
}
