% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Small.R
\name{lm_eqn}
\alias{lm_eqn}
\title{lable linear regression formula}
\usage{
lm_eqn(df, Dep, Exp)
}
\arguments{
\item{df}{A data.frame object contains two variables: dependent variable (Dep) and explanation variable (Exp).}

\item{Dep}{A character specify dependent variable. The first column will be dependent variable as default.}

\item{Exp}{A character specify explanation variable. The second column will be explanation variable as default.}
}
\value{
a linear regression formula
}
\description{
lable linear regression formula
}

