
\name{motif.enrichment.plot}
\alias{motif.enrichment.plot}
\title{motif.enrichment.plot to plot bar plots showing motif enrichment ORs and 
95\% confidence interval for ORs}
\usage{
motif.enrichment.plot(motif.enrichment, significant = NULL, dir.out = "./",
  save = TRUE, label = NULL)
}
\arguments{
\item{motif.enrichment}{A data frame or a file path of get.enriched.motif output
motif.enrichment.csv file.}

\item{significant}{A list to select subset of motif. Default is NULL.}

\item{dir.out}{A path specifies the directory to which the figures will be saved.
Current directory is default.}

\item{save}{A logic. If true (default), figure will be saved to dir.out.}

\item{label}{A character labels the output figures.}
}
\value{
A figure shows the enrichment level for selected motifs.
}
\description{
motif.enrichment.plot to plot bar plots showing motif enrichment ORs and 
95\% confidence interval for ORs. Option motif.enrichment can be a data frame 
generated by \code{\link{get.enriched.motif}} or a path of XX.csv saved by the 
same function.
}
\references{
Yao L, Shen H, Laird PW, Farnham PJ,Berman BP: Inferring Regulatory Element 
Landscapes and Transcription Factor Networks from Cancer Methylomes. in revision of
Genome Biology
}
\author{
Lijing Yao (maintainer: lijingya@usc.edu)
}

\examples{
motif.enrichment <- data.frame(motif=c("TP53","NR3C1","E2F1","EBF1","RFX5",
"ZNF143", "CTCF"),
OR=c(19.33,4.83,1, 4.18, 3.67,3.03,2.49),
lowerOR =c(10,3,1.09,1.9,1.5,1.5, 0.82),
upperOR =c(23,5,3,7,6,5,5),
stringsAsFactors=FALSE)
motif.enrichment.plot(motif.enrichment=motif.enrichment,
                     significant=list(OR=3),
                     label="hypo", save=FALSE)
}

