
\docType{class}
\name{Pair-class}
\alias{Pair-class}
\title{An S4 class that pairs information, probe information and gene information.}
\description{
An S4 class that pairs information, probe information and gene information.
}
\section{Slots}{

\describe{
\item{\code{pairInfo}}{A data.frame}

\item{\code{probeInfo}}{A GRanges object.}

\item{\code{geneInfo}}{A GRanges object.}
}}

\seealso{
  \code{\link{schematic.plot}}, \code{\link{getGeneInfo}}, \code{\link{getProbeInfo}},
  \code{\link{fetch.pair}}
}
\author{
Lijing Yao (maintainer: lijingya@usc.edu)
}
