% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{Provider}
\alias{Provider}
\title{A ChatBot Provider}
\usage{
Provider(
  base_url = class_missing,
  api_key = class_missing,
  extra_args = class_missing
)
}
\arguments{
\item{base_url}{The base URL for the API.}

\item{api_key}{The api key used for authentication.}

\item{extra_args}{Arbitrary extra arguments included in the request body.}
}
\description{
A provider captures the details of one chatbot service/API. To add support
for a new backend, you will need to subclass \code{Provider} (adding any
additional fields that your provider needs) and then implement the
various generics that control the behaviour of each provider.
}
