% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scatter.plot.R
\name{scatter}
\alias{scatter}
\title{scatter}
\usage{
scatter(meth, exp, legend.title = "Legend", category = NULL,
  xlab = NULL, ylab = NULL, ylim = NULL, dots.size = 0.9,
  title = NULL, correlation = FALSE, color.value = NULL,
  lm_line = FALSE)
}
\arguments{
\item{meth}{A vector of number.}

\item{exp}{A vector of number or matrix with sample in column and gene in rows.}

\item{legend.title}{Plot legend title}

\item{category}{A vector of sample labels.}

\item{xlab}{A character specify the title of x axis.}

\item{ylab}{A character specify the title of y axis.}

\item{ylim}{y-axis limit i.e. c(0,25)}

\item{dots.size}{Control dots size}

\item{title}{A character specify the figure title.}

\item{correlation}{Show person correlation values}

\item{color.value}{A vector specify the color of each category, such as}

\item{lm_line}{A logic. If it is TRUE, regression line will be added to the graph.}
}
\value{
A ggplot figure object
}
\description{
scatter
}
