% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schematic.plot.R
\name{schematic.plot}
\alias{schematic.plot}
\title{schematic.plot to plot schematic plots showing the locations of genes and probes.}
\usage{
schematic.plot(pair, 
               byProbe, 
               byGene, 
               byCoordinate = list(chr = c(), start = c(), end = c()), 
               dir.out = "./", save = TRUE, ...)
}
\arguments{
\item{group.col}{A column defining the groups of the sample. You can view the 
available columns using: colnames(MultiAssayExperiment::pData(data)).}

\item{pair}{A Pair object. All slots of Pair class should be included}

\item{byProbe}{A vector of probe names.}

\item{byCoordinate}{A list contains chr, start and end. 
byCoordinate=list(chr=c(),start=c(),end=c()).}

\item{dir.out}{A path specify the directory for outputs. Default is current directory}

\item{save}{A logic. If true, figures will be saved to dir.out.}

\item{...}{Parameters for GetNearGenes}

\item{byGene}{A vector of gene ID}
}
\description{
schematic.plot is a function to plot schematic plots showing the locations of genes and probes.
}
\details{
byProbes:
When a vector of probes' name are provided,
function will produce schematic plots for each individual probes.
The schematic plot contains probe, nearby 20 (or the number of gene user specified.)
genes and the significantly linked gene to the probe.

byGene:
When a vector of gene ID are provided, function will produce schematic plots 
for each individual genes. The schematic plot contains the gene and all the 
significantly linked probes.

byCoordinate:
When a genomic coordinate is provided, function will 
produce a schematic plot for this coordinate. The schematic plot contains 
all genes and significantly linked probes in the range and the significant links.
}
\examples{
data(elmer.data.example)
pair <- data.frame(Probe = c("cg19403323","cg19403323", "cg26403223"),
                   GeneID = c("ENSG00000196878", "ENSG00000009790", "ENSG00000009790" ),
                   Symbol = c("TRAF3IP3","LAMB3","LAMB3"), 
                   Pe = c(0.001,0.00001,0.001))
new.schematic.plot(data, pair, byProbe = "cg19403323")
new.schematic.plot(data, pair, byGeneID = "ENSG00000009790")
new.schematic.plot(data, pair, byCoordinate = list(chr="chr1", start = 209000000, end = 209960000))
}

