% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Schematic.plot.R
\name{schematic}
\alias{schematic}
\title{schematicPlot}
\usage{
schematic(probe.range, gene.range, special = list(names = c(), colors = c()),
  interaction = list(probe = c(), gene = c(), colors = c()), label,
  save = TRUE)
}
\arguments{
\item{probe.range}{A GRanges object contains probes coordinates.}

\item{gene.range}{A GRanges object contains gene TSS coordinates.}

\item{special}{A list: special=list(names=c(), colors=c()) show the name of
feature you want to highlight and specify the color respectively.}

\item{interaction}{A list: interaction=list(probe=c(),gene=c(),colors=c())
show the interacted features and specify the color respectively.}

\item{label}{A character labels the outputs figure.}

\item{save}{A logic. If true, figure will be saved to dir.out.}
}
\value{
A schematic plot.
}
\description{
schematicPlot
}
\details{
byProbes When a vector of probes' name are provided, function will
produce schematic plot for each individual probes. The schematic plot contains
probe, nearby 20 (or the number of gene user specified.) genes and the significantly
linked gene with the probe.
}

